\name{sOG-package}
\alias{sOG-package}
\alias{sOG-package}
\docType{package}
\title{Removing the influence of group variables
}
\description{
\packageDescription{sOG}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sOG}
\packageIndices{sOG}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sOG}

Maintainer: \packageMaintainer{sOG}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
