% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOG.R
\name{sOG}
\alias{sOG}
\title{sparse Orthogonal to Groups analysis}
\usage{
sOG(X = warning("X missing"), z = stop("z missing"),
  t_val = stop("t_val missing"), K = max(2, round(NCOL(X)/100)),
  rescale = T, control = control_sOG())
}
\arguments{
\item{X}{n x p data matrix to preprocess.}

\item{z}{n x 1 vector with group information.}

\item{t_val}{penalty parameter. Must be a positive number. Control the strength of the l-1 penalty.}

\item{K}{approximation rank. The defaul is \code{max(2, round(NCOL(X)/100))}}

\item{control}{Control parameters for the fitting process. A list as returned by \code{\link[sOG]{control_sOG}}.}

\item{rescaled}{Should the matrix \code{X} be rescaled? Default is \code{TRUE}.}
}
\value{
\itemize{
\item \verb{U} n x K matrix of scores.
\item \verb{S} K x p matrix of sparse loadings
}
}
\description{
Perform sparse matrix decomposition under group constraints.
}
\details{
The function performs a sparse matrix decomposition of the input matrix \code{X} with constraints on the left singular vectors and the group variable \code{z}.
The output is a matrix \code{U} of basis and a matrix \code{S} of scores such that \eqn{\tilde X = S * U^T}, where \eqn{\tilde X} is the rank-K approximation of \code{X}
}
\examples{
k.rid = 10
n = 5000
p = 200
W = matrix(rnorm(p*k.rid), k.rid)
S = matrix(rnorm(n*k.rid), n)
z = sample(rep(0:1, each=n/2))
lambda = rnorm( k.rid, mean = 0, sd = 1)
A = jitter( (S - lambda * z ) \%*\% W)
res = sOG(X = A, z = z, t_val = 10, K = 30)
}
\references{
Aliverti, Lum, Johndrow and Dunson (2018). Removing the influence of a group variable in high-dimensional predictive modelling (https://arxiv.org/abs/1810.08255).
}
