\name{sOG-package}
\alias{sOG-package}
\alias{sOG-package}
\docType{package}
\title{Removing the influence of group variables }
\description{ \packageDescription{sOG}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{sOG}
\packageIndices{sOG}
}
\author{
\packageAuthor{sOG}
Maintainer: \packageMaintainer{sOG}
}
\references{
Aliverti, Lum, Johndrow and Dunson (2018). Removing the influence of a group variable in high-dimensional predictive modelling (https://arxiv.org/abs/1810.08255).
}
\keyword{ package }
\seealso{
}
\examples{
}
